+++ pkgs/development/python-modules/nose/default.nix
@@ -1,10 +1,10 @@
-{ lib
-, buildPythonPackage
-, fetchPypi
-, python
-, coverage
+{
+  lib,
+  buildPythonPackage,
+  fetchPypi,
+  python,
+  coverage,
 }:
-
 buildPythonPackage rec {
   version = "1.3.7";
   pname = "nose";
@@ -27,20 +27,25 @@ buildPythonPackage rec {
     2to3 -wn nose functional_tests unit_tests
   '';
 
-  propagatedBuildInputs = [ coverage ];
+  propagatedBuildInputs = [coverage];
 
-  doCheck = false;  # lot's of transient errors, too much hassle
-  checkPhase = if python.is_py3k or false then ''
-    ${python}/bin/${python.executable} setup.py build_tests
-  '' else "" + ''
-    rm functional_tests/test_multiprocessing/test_concurrent_shared.py* # see https://github.com/nose-devs/nose/commit/226bc671c73643887b36b8467b34ad485c2df062
-    ${python}/bin/${python.executable} selftest.py
-  '';
+  doCheck = false; # lot's of transient errors, too much hassle
+  checkPhase =
+    if python.is_py3k or false
+    then
+      ''
+        ${python}/bin/${python.executable} setup.py build_tests
+      ''
+    else
+      ""
+      + ''
+        rm functional_tests/test_multiprocessing/test_concurrent_shared.py* # see https://github.com/nose-devs/nose/commit/226bc671c73643887b36b8467b34ad485c2df062
+        ${python}/bin/${python.executable} selftest.py
+      '';
 
   meta = with lib; {
     description = "A unittest-based testing framework for python that makes writing and running tests easier";
     homepage = "http://readthedocs.org/docs/nose/";
     license = licenses.lgpl3;
   };
-
 }
